--- src/client/datascience/jupyter/jupyterKernelPromiseFailedError.ts
@@ -1,9 +1,0 @@
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT License.
-'use strict';
-
-export class JupyterKernelPromiseFailedError extends Error {
-    constructor(message: string) {
-        super(message);
-    }
-}
